requirements.txt
# API Client
python-binance>=1.0.19

# Database
SQLAlchemy>=2.0.0

# WebSocket
websockets>=12.0

# Environment Variables
python-dotenv>=1.0.0

# Logging
structlog>=24.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Type Checking
mypy>=1.8.0
types-python-dateutil>=2.8.19.14

# Code Quality
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0 